+++ src/dtos/user-saved-post.dto.ts
@@ -1,9 +1,1 @@
-import { ApiProperty } from '@nestjs/swagger';
 
-export class SavePostResponseDTO {
-  @ApiProperty({ example: 201 })
-  statusCode: number;
-
-  @ApiProperty({ example: 'Post saved successfully.' })
-  message: string;
-}
